requirements.txt
pyyaml>=6.0
# For high-quality figure export (optional):
# pip install playwright && playwright install chromium
